+++ src/main/java/ControlledMoney/api/domain/conta/dtos/ContaInputDTO.java
@@ -8,7 +8,7 @@ public record ContaInputDTO(
     Long idUsuario,
 
     @NotNull
-    Long saldo
+    Double saldo
 ) 
 {
